refactor: rate 패키지 분리 (#89)
