--- src/main/java/com/gugucon/shopping/item/dto/request/RateCreateRequest.java
@@ -1,16 +1,0 @@
-package com.gugucon.shopping.item.dto.request;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor
-public class RateCreateRequest {
-
-    private long orderItemId;
-
-    private short score;
-}
