+++ src/main/java/com/gugucon/shopping/rate/dto/response/RateResponse.java
@@ -1,0 +1,15 @@
+package com.gugucon.shopping.rate.dto.response;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class RateResponse {
+
+    private int rateCount;
+    private double averageRate;
+}
